fd593a3 COMP: Removed massive amounts of warnings, changed boolean defaults
bc95c29 ENH: Adding ToString to sitkExceptionObject
b785857 COMP: Fixing warnings about mis-matched parameters to ITK methods
8baafff BUG: fixing unsigned values across languages
209f300 COMP: Fixing compile warnings
6644a8f BUG: method was not using parameter is was suppose to
714a325 COMP: marked unused parameters as unsused
42c0d73 WARN: Removed warnings in registration code
bf7d2db WARN: Mismatch between signed and unsigned
b80b2a1 WARN: Removed warnings from Image class

+++ Testing/Unit/BasicFiltersTests.cxx
@@ -15,9 +15,9 @@
 TEST(BasicFilters,RecursiveGaussian_ENUMCHECK) {
 
   typedef itk::RecursiveGaussianImageFilter< itk::Image<float,3> > ITKRecursiveGausianType;
-  EXPECT_EQ( ITKRecursiveGausianType::ZeroOrder, itk::simple::RecursiveGaussianImageFilter::ZeroOrder );
-  EXPECT_EQ( ITKRecursiveGausianType::FirstOrder, itk::simple::RecursiveGaussianImageFilter::FirstOrder );
-  EXPECT_EQ( ITKRecursiveGausianType::SecondOrder, itk::simple::RecursiveGaussianImageFilter::SecondOrder );
+  EXPECT_EQ( (int)ITKRecursiveGausianType::ZeroOrder, (int)itk::simple::RecursiveGaussianImageFilter::ZeroOrder );
+  EXPECT_EQ( (int)ITKRecursiveGausianType::FirstOrder, (int)itk::simple::RecursiveGaussianImageFilter::FirstOrder );
+  EXPECT_EQ( (int)ITKRecursiveGausianType::SecondOrder, (int)itk::simple::RecursiveGaussianImageFilter::SecondOrder );
 }
 
 TEST(BasicFilters,Cast) {
